refactor: make one command for update
